nfs: move v2 parsing into own file
