+++ rust/src/nfs/mod.rs
@@ -22,6 +22,7 @@ pub mod nfs2_records;
 pub mod nfs3_records;
 pub mod nfs4_records;
 pub mod nfs;
+pub mod nfs2;
 pub mod nfs3;
 pub mod nfs4;
 pub mod log;
